Update README.md
fixes typo... I am new to open source contributions, and looking for a project. I stumbled upon yours and noticed this small typo. Not sure if a PR is the right way to bring it to attention, but if not, please excuse the annoyance! Just a bit of practice for me to learn the process!